pyproject.toml
#[[tool.poetry.source]]
#name = "pypi"
#url = "https://pypi.example.org/simple/"
#priority = "primary"

[tool.poetry]
name = "FlightRadar"
version = "0.0.1"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "*"
hypercorn = "*"
typer = "*"
requests = "*"
tenacity = "*"
pydantic = "*"
pydantic-settings = "*"
asyncpg = "*"
sqlalchemy = "*"
psycopg2 = "*"
tortoise-orm = "*"
aerich = "*"
python-dotenv = "*"
aiohttp = "*"

[tool.poetry.dev-dependencies]
requests-mock = "*"
black = "*"
pytest = "*"
pytest-sugar = "*"
pytest-asyncio = "*"
pytest-docker = "*"
pytest-mock = "*"
pytest-docker-compose = "*"
pytest-env = "*"

[tool.poetry.scripts]
makemigrations = "aerich migrate"
migrate = "aerich upgrade"

[tool.aerich]
tortoise_orm = "setup.AERICH_DB_CONFIG"
location = "./migrations"
src_folder = "./."
[requires]
python_version = "3.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"